pyproject.toml
[tool.poetry]
name = "pancaketrade"
version = "0.6.1"
description = "Trading bot for PancakeSwap"
authors = ["Valentin Bersier <vbersier@gmail.com>"]

[tool.poetry.scripts]
trade = 'pancaketrade:main'

[tool.poetry.dependencies]
python = ">3.7.1,<3.10"
requests = "^2.25.1"
importlib-metadata = {version="^4.0.1", python="<3.8"}
loguru = "^0.5.3"
python-telegram-bot = "^13.4.1"
web3 = "^5.18.0"
cachetools = "^4.2.2"
PyYAML = "^5.4.1"
click = "^7.1.2"
yamale = "^3.0.6"
questionary = "^1.9.0"
peewee = "^3.14.4"
apscheduler = "<4.0.0"
python-nubia = { file = "python_nubia-0.3.0-py3-none-any.whl"}

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"
black = { version = "*", allow-prereleases = true }
mypy = "<0.900"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
